Fixed RavenDB-328: Query error now disappears when retrying query.
Re-organised VirtualCollection code a little. VirtualCollection is no longer responsible for handling errors that occur in data-fetch tasks: VirtualCollectionSource handles these in coopreation with the parent viewmodels.

Enabled VirtualCollectionSource to keep track of when it is busy

VirtualCollection now nulls out any data items in pages that it failed to fetch.