Add full support for facet searches.
Use a collector on the actual Lucene query to get all of the terms
resulting from the search and count the hits along the way.  Now we can
get all term hits without extra queries.  Also add support for limiting
the number of returned facet values in the facet setup document and add a
sorting mode for the selection of those N facet values.