Use logging abstraction in Raven.Datbase
- Rename ILogProvider \ LogProvider to ILogManager \ LogManager respectively.
This matches NLog usage, but if it leads to confusion, can be reverted.

- Added IsDebugEnabled and IsWarnEnabled to ILog. I'd prefer if the deferred
log message creation was used (passing Func<> paramater) but I didn't want to
change some of the existing logic around IsDebugEnabled. IsInfoEnabled etc can
be added as required.

- Fixes a load of places that have a "using Nlog;" directive.

- This commit contains a compilation failure as there is a feature, the Logs.cs
request responder, that directly depends on NLog for reasons other than
logging. This will be fixed in a subsequent commit.

Signed-off-by: Damian Hickey <dhickey@gmail.com>