RavenDB-665 When specifing the Result type in the index, the compiler automatically omits any cast to the type. This works: public class NullableIndex : AbstractMultiMapIndexCreationTask<Result> 		{ 			private class Result : Nullable 			{ 			}
			public NullableIndex()
			{
				AddMap<Nullable>(nullables => nullables.Select(nullable => new
				{
					String = (string)null,
					Decimal = (decimal?)null,
					Double = (double?)null,
					Float = (float?)null,
					Long = (long?)null,
					Int = (int?)null,
					Short = (short?)null,
					Byte = (byte?)null,
					Bool = (bool?)null,
				}));
			}
		}

But this does not:
public class NullableIndex : AbstractMultiMapIndexCreationTask<Result>
		{
			private class Result : Nullable
			{
			}

			public NullableIndex()
			{
				AddMap<Nullable>(nullables => nullables.Select(nullable => new Result
				{
					String = (string)null,
					Decimal = (decimal?)null,
					Double = (double?)null,
					Float = (float?)null,
					Long = (long?)null,
					Int = (int?)null,
					Short = (short?)null,
					Byte = (byte?)null,
					Bool = (bool?)null,
				}));
			}
		}

The issue is that the compiler will strip the cast (string) since the type is know at compile time. The issue that the type is omitted in the client before it sent to the server.

Starting from the next build, the following will be supported too:
public class NullableIndex : AbstractMultiMapIndexCreationTask<Result>
		{
			private class Result : Nullable
			{
			}

			public NullableIndex()
			{
				AddMap<Nullable>(nullables => nullables.Select(nullable => new Result
				{
					String = null,
					Decimal = null,
					Double = null,
					Float = null,
					Long = null,
					Int = null,
					Short = null,
					Byte = null,
					Bool = null,
				}));
			}
		}

We automatically infer the type and cast the null to it type. An effort was done to use CSharp keywords for the type instead of the full .net framework name, in order to keep the index definition as small as possible.