Allows IndexedProperties Bundle to update complex properties
Current implementation updates complex properties as a string instead of
an object. For example: if I have an index which creates a Rating for
each Article (an aggregate of Votes).  When this is updated by the
IndexedProperty Bundle what I get is:

"Rating": "{\"Score\"::40754086, \"Count\":111, \"VoteTotal\":1839,
\"WeightTotal\":598, \"VoteCounts\":[22,19,24,18,28]}"

When it should be

"Rating": {"Score"::40754086, "Count":111, "VoteTotal":1839,
"WeightTotal":598, "VoteCounts":[22,19,24,18,28]}