RavenDB-1033 Fixing the problem that streaming was not working on IIS when compression was enabled. The problem was the usage of:
response.BufferOutput = false and Request.Filter  together. The zLib exception was thrown but only when the request was closing. The applied fix allows to stream the response however at the end of the response method we have to turn off the buffering to avoid response flushing and writing to already closed compression stream.