Facet bug, query bug, loadstartingwith case bug
Raven bugs:

creating a facet hides other errors.

facet doesn’t work with datetimes and the advanced api

authorization doesn’t work with transformers

allowing for expressions when creating facet ranges with advanced api
|| in where blows chunks.

When we specify our own Id with mixed case, we can't find it with an all
lowercase prefix request nor with the matching mixed-case request.
LoadStartingWith is partially case-sensitive.. And Load is not?