If you do
LoadByUniqueConstraint<Foo>(arrayOfIds);

foreach(var id in arrayOfIds) {
    LoadByUniqueConstraint<Foo>(id);
}
You get a request for every Foo that does not exist. Using a Guid as the "id" in this case is a useful hack that avoids having to line up 2 arrays, but there is no need to generate a new Guid each time. Instead I've just hardcoded a fixed Guid and we'll use this every time.

Manually patch from Gareth Thackeray's PR