Created the setting "Raven/ImplicitFieldsToFetchFromDocumentMode"
This setting should be used to control how we handle fields that we want
projectet, but are missing from the index.
Currently we're loading the document itself to retrieve the missing
values, but with this setting we will be able to configure what RavenDB
does in this situation:
1 Get the values from the document
2 Do not fill those values
3 Throw an exception

I'm currently not sure how the third case will work out, but we'll see
once I've implemented the usage of this setting in the DocumentRetriever
class.