All the columns names are being saved in order to be able to show them autocratically when pressing the bind field in choose columns button. At the viewModel - added a new function getColumnsNames() which returns all the columnsNames Now,selectColumns is being called with all the columns names and all the list of columns names appear autocratically when inserting a new row at the choose columns function.
The columns names need to updated all the time, following changes like - deleting a row deleting the binding field.
In order to keep the name updated I've added the method - updateColumnsNames which checks if columns names appear in the custom columns array and update autoCompleteBase with
the new updated array without names which already appear in customColumns.

When monitorForNewRows is being called it subscribe the binding column for updates in columns names - in case someone deletes just the column name in the binding field,
and also calls for updateColumnsNames when subscribing the columns so it will show updated results when deleting the entire row.

For query - selectColumns works the same way- it receives all the columns names
In order to get all the columns names I created getQueryGrid which return virtualTable

Signed-off-by: Hila Mefano <hila@ayende.com>