Added GetUserInfo - a method which enables to get all the information of the user- databases, readOnlyDatabases, readWriteDatabases admin permissions etc, Added GetUserPermission- this method receives database name and MethodOption- PUT/GET and checks if the database name appears in the databases list (retrieved by GetUserInfo). If the database dosen't exists the user has no permissions. if it does exists then we check the permissions - admin / ReadOnly. if ReadOnly isn't set to true then the user has ReadWrite Permission This function returns explanation on whatever the user is granted this permission or not, and why.
Added UserPermission class with isGranted and Reason properties for the use of GetUserPermission - permission is granted or not with a reason why.

In a case of OAuthPrincipal - the fields: AdminDatabases, ReadOnlyDatabases and ReadWriteDatabases weren't set to any value.
I'm setting these fields in the constructor of OAuthPrincipal.
Also DatabaseInfo had no property which indicates the the database is readOnly. I added isReadOnly in DatabaseInfo and when setting the Databases fields in case of OAuthPrincipal, this field is also being set.

added the URL for a permission request at the server side

Signed-off-by: Hila Mefano <hila@ayende.com>