When calling ImportData there was an issue catching the exception being thrown because ImportData IS A Task. The changes I made in order to be able to catch the exception type without loosing it:
change exportSectionRegistar Dictionary to type: Dictionary<string, Func<Task<int>>>

when calling exportSectionRegistar.Add for "Indexes": Add now receives Func<Task<int>>
and each time we call ImportIndexes we don't run it using Results.
changes:          var indexCount = ImportIndexes(jsonReader).Result
to
 var indexCount = await ImportIndexes(jsonReader).ConfigureAwait(false);

we use await on the ImportIndexes

same for exportSectionRegistar.Add for "Docs" is also changed to be async and we await on ImportDocuments

same for exportSectionRegistar.Add for "Attachments" is also changed to be async and we await on ImportAttachments

and same for exportSectionRegistar.Add for "Identities" is also changed to be async and we await on ImportIdentities

All these changes are needed because ImportData is Task and we dont want to loose the exception type.

changed: currentAction to be a Func<Task<int>> beacuse we are trying to get the value of exportSectionRegistar which is of the same type:
exportSectionRegistar.TryGetValue(currentSection, out currentAction)
and therefore we also doing await on  currentAction() in order to set exportCounts[currentSection]