RaveDB-3979-  Studio- added a checkbox - Disable versioning during import at the import file settings.
Raven.Database.FileSystem.Controllers -
added at catch(Exception e) a check for OperationVetoedException
when this exception accrues we inform the client to disable versioning during import

Versioning/VersioningUtil.cs-IsVersioningDisabledForImport- added a a method IsVersioningDisabledForImport which checks if the metadata contains RavenIgnoreVersioning and if its value is set to true, so we can check if the versioning bundle with disabled during import

Raven.Database\FileSystem\Bundles\Versioning\Plugins\VersioningTriggerActions - added checks when AllowPut is called with a null file we should allow put only if versioining isn't active or if versioning during import is disabled
If changes to revisions aren't allowed and the metadata contains "Historical" we Deny put.
Also, splited the check first checking if IsVersioningActive is true then if IsVersioningDisabledForImport is true as well then vetoResult is Allowed. else we continue to check for revisions like before

AsyncFilesServerClient.cs - if the exception is OperationVetoedException we throw the exception with its message, this is needed here because otherwise we loose the exception type.

RavenDB_3979 - added a test for importing into a file system with versioning disabled and enabled.
When versioning is enabled the import fails and an exception is thrown, informing the client to disable versioning,
When versiioning is disabled the import succeeds.

Fixed- tabs to spaces

RavenDB-3736 -
Raven.Database/Server/Controllers/BulkInsertController.cs - when OperationVetoedException is thrown, we need to throw the message.

RavenDB-3736 - Added a test for importing into a database when versioning is disabled and enabled.
When the ShouldDisableVersioning flag is set to false the importing fails and the client is informed to set the flag to true before importing.
When the ShouldDisableVersioning flag is set to true the import succeeds.

Corrections in tests:

Raven.Tests.Bundles/Versioning/Versioning.cs - fixed the test- the flag ShouldDisableVersioningBundle needs to be set to true, for the operation of import to succeed.

Raven.Tests.FileSystem/Bundles/Versioning/VersioningTests.cs - at CannotModifyHistoricalRevisionByDefault_PUT- the Exception type that's being thrown needs to be OperationVetoedException