RavenDB-4008: Translation Table
Implements the layout of the chunks translation table. Every single chunk will host a cache friendly representation of a subtree of the whole-tree. The mapping between the node-name (which is the number of the node in a complete binary tree) into a chunk and it's relative name in the subtree is being handled by the TranslationTableHeader which stores in a list of longs those maps implicitly. Chunk 0 represents the tree where the root of the tree reside. Chunk 1 holds the subtree created by the left-most node (at depth D) in the Chunk 0 subtree and so on.

Fixed the BitVector constructors to build properly prefix-free bit vectors with different .Of constructors.

Cherrypicked from: 35a9a2841c81cc9bf21a0fcbfc5513d25d8dcc51 [35a9a28]