RavenDB-4008: Complete implementation for Delete
Implemented the Delete operation for Prefix Trees. It passes structural tests when running all changes in the same transaction. Changes have to be made to ensure that there are not problems with multiple transactions.
The default now is the subtree depth to ensure that we load a whole L1 cache line per node chunk. (this parameter can be tweaked for better performance, right now we havent done any performance work on this).