RavenDB-4333: ByteString implementation
The new ByteString will allow to allocate everything from unmanaged buffers and kill all the memory in one go or release memory and being reused.
It will also allow to create external pointers to memory, while still retaining the ability to allocate them very fast.
The ByteString will be allocated by the ByteStringContext and for the caller code purpose it behaves like a pointer using some struct/JIT magic.

RavenDB-4568: Slice are not pointers

Exploiting the implementation of the ByteString we could create Slices that are entirely allocated on the stack and whose size is exactly the same to a reference pointer.
The side effect is that Slices are not very fast to pass-around, but they will also create aliases (the older slices didn't) which is a breaking change.

(+11 squashed commits)