RavenDB-4866: Performance degradation is being caused by memory fragmentation.
The current ByteStringContext change will try to diminish the fragmentation keeping track of a global pool. While potentially complicated to fix the issue for good, this is an initial approximation (good enough) to deal with the issue. More strict long term performance profiling should be performed to ensure that this strategy can lower fragmentation of the unmanaged heap to the point we need it.