Ensuring tokenizers hot-code gets inlined avoiding virtual method calls.
Instead of implementing the hot methods as virtual overrides, we instead introduce a generic base class which allows to specialize using an struct implementing the ILowerCaseTokenizerHelper. This allow us to implement each virtual method isolated from the general algorithm that gets executed and the JIT can pickup the code and inline and optimize it accordingly.