Improvements for StringSegment, ReadNumber and ProcessTokenOffsetFlags
We changed the order of the parameters of the constructors of StringSegment in order to be able to exploit the knowledge and avoid branching in the code layout.
ReadNumber now sports a tighter code layout and it is being able to inline in several more call-sites.
ProcessTokenOffsetFlags and similar code used a switch statement which does not exploit the probability of execution of each branch. While the switch would be converted into if-then-else statements at the JIT level, the order in which the branches occur is suboptimal. Instead we translated it manually and ordered it accordingly of the knowledge on how common each one is.