Improvements to the FastDictionary that allow us to inline usually virtual calls.
The FastDictionary follows the same idea of Dictionary of supporting IEqualityComparer based comparers in order to play along with the framework. We dont want to lose that, but at the same time we can allow us to have that functionality baked in using JIT magic. We use a similar trick to the ones we use to override the virtual dispatching using JIT ability to specialize the code when faced with an struct and also add the ability of the JIT to devirtualize the calls for more standard not so optimized code on CoreCLR 2.0 implemented by Andy Ayers at Microsoft.