RavenDB-6629: Fixed degradation of dictionary performance because of a not good enough hashing function.
Implemented Marvin32 because we needed a compact function to allow inlining of key methods. xxHash while very fast has a big instructions footprint making it a very unsuitable hash function on hot-paths and small inputs. While we pondered other options like Dbj or FNV which are very fast for small inputs, the data we are going to be stores comes unfiltered from user land; so for a very small premium we decided to use a flood resistant one.

We also implemented improvements to the code layout for the FastDictionary insertion method.