RavenDB-6827 Discover server address by scanning the range of possible IPs and comparing server ID
To avoid the need of escalated privileges (for running HyperV commands) for running this script we omit this problem by scanning IP range to find the server. This is going to help people having customized DockerNAT network address. When using defaults - 10.0.75.2 then only 1 call is going to be made to the server and this covers most of the cases. In other cases we're scanning up to 253 addresses. Added `-DontScanVmSubnet` flag which might be used to skip the scan altogether and fallback to displaying 10.0.75.2.