RavenDB-10375 We need to use the commit charge, rather than the memory utilization, for low memory
We consider low memory only if we don't have enough free pyhsical memory *AND* the commited memory size if larger than our pyhsical memory.
This is to ensure that from one hand we don't hit the disk to do page faults and from the other hand
we don't want to stay in low memory due to retained memory.