RavenDB-10830 Fixing "Could not set the size of file compression.0000000000.buffers to ...  ---> System.ComponentModel.Win32Exception: The requested operation cannot be performed on a file with a user-mapped section open"
The issue was that we failed on EnsureContinuous call (InsufficientMemoryException was thrown) but we already managed to increase the size of the backing file. Afterwards next attempt to reduce the size of physical file requires all readers of MMF to stop using it (close) before we increase it. The solution is to create next temp buffer file.