RavenDB-11354 High CPU after hitting leader stepdown several times
The issue is related to code for sending topology change notification from the leader.
We waited to leave 'Follower' state in order to reconnect to a new leader, but it could happend that the leader was changed without us leaving the follower state, and so we were connected to the wrong node.
Ihis could caused a circular connection between the nodes which sent the same notification again and again.