RavenDB-11366 InvalidDataException Expected to get type of 'LogLengthNegotiationResponse' message, but got 'Error' message.
The issue is probably due to a race between disposing a follower in term X, while creating a new one in the same term.
This could occur during a short period split-brain (network delay) so the follower went to elections but was immediately forced by the leader to becamce his follower again.