RavenDB-11015 Verify cluster behavior on node removal
Addressing a race when rejoining a removed node.
In case that the cluster hasn't truncated its log, rejoinig the node would result in a different state on the nodes.
Since the leader wouldn't send 'install snapshot' to override the existing state.
But instead it would send the missing entries to the node.