RavenDB-11892 Memory leak in ByteStringContext
Resetting the context doesn't free the memory as long as we had only 2 segments.

So with many small allocation, this caused to the hot segments to grow to their maximal value of 16 MB.
And in presence of many contexts we ended up in a high memory usage.

Since ByteStringContext is mostly used for converting from string to slice document ids and different keys for voron tress/indexes, we do not expect to have large segments.

Therefor we changed the default size of minimum block from 64k to 4k and the maximal segment size to 2MB.