RavenDB-11923 Smuggler.LegacySmugglerTests.CanImportIndexesAndTransformers
When running in-memory all of the temp data for indexes were created in the same temp folder, which in this case genreated over 1,500 files in a single folder.
This crippled the database deletion, since it had to enumerate so many files (`Directory.GetFiles(tempPath.FullPath, "*.buffers")`).

So the fix is to create each temp data of the index in a seperate folder.