RavenDB-11968 Fix CanPreformSeveralClusterTransactions test that throws NotLeadingException
The node could have been the leader while the rachis request arrived, but lost it afterwards.
In that case we need to throw the appropriate exception of the client (the executor that issued the request).
Moreover, this fix will cause for the raft command to be retried as long as we have a any valid leader.