RavenDB-11352 Pull Replication
In pull replication there are two roles: central and edge.

In the central cluster we define and keep on the database record a pull replication task with all the relevant settings like mentor, delay and certificates.
Once defined the central will wait for a connection to be initiated by the edge.

On the edge node we define an external replication with pull replication settings.
This means that a node in an edge cluster will request the tcp info for some node in the central and initiate the TCP connection so the central will send the document via this connection.

Security:
The definition on the central will contain thumbprints for allowed certificates.
The edge need to put a certificate (with the private key) that corresponds to the thumbprint of the central definition. This certificate is store with different prefix then the regular client certificates so we don't trust it by mistake.
Initiating a pull replication will use the edge certificate which the central should allow using the stored thumbprint in the corresponded definition.