RavenDB-11871 Journal applicator keep applying journals even if the indexes didn't change for hours
The issue was that we tried to create journals that are power of 2, but if the transaction size was greater than the size of the journal, we would create a journal file that was bigger than the actual transaction size.
That, in turn, would lead to having available space in the journal file, so we would keep it around after sync.
This change is simply to ensure that the size of the journal file will be either the max journal size or the space the transaction requires. In this case, after a large transaction, we'll no longer have any space there and can clear the large journal file.
As a result, on recovery, we'll not have any journal files to apply and can reduce the speed we take to startup