RavenDB-12832 keep conflicted revision with its original change vector
resolved conflict kept the conflicted revision with a new change vector (locally modified), so it could happened that multiple nodes resolved the conflict locally and replicated them to one another.

bonus: same code path for a revision or a deleted revision with higher change vector to overwrite the existing document.