RavenDB-13181 Counters replication discrepancies
The underlying issue is that we resotred not in the insertion order in which the document, revisions, coutners, etc. were putted in the original database.
In the restored database this could cause that a document with higher change vector will have a lower local etag a therefore will skip the replication of doucments with lower change vector.

- Remove keep original change vector
- Insert the counters by insertion order