RavenDB-13612 follower should report back the latest incoming entry index
The underlaying issue was that the follower reported back his latest entry, which might be from a different term.
This would cause the leader and the follower to commit different entries on the same index.

another cause that incerased the probability to this issue is that the leader started to negotiate before appending 'noop' to his own log.
that might lead to a more up-to-date follower not to rewind his log.