RavenDB-14044 Error in ClusterMaintenanceSupervisor
The root cause to this issue is that the supervisor, that collects status info from other nodes didn't handle properly the case when the there is more than one status report in the network buffer.
In that case, it parsed the first report and abandoned the part of the second report, so in the next iteration the parsing of the second report failed.