RavenDB-16084 GetConflictStatus should return correct result
As far as I can tell this affects databases that has 'UnusedIds' together with cluster tx.

Consider the case where our local database cv is `A:11, B:10` and node C was remvoed.
So if we compare it with a doc that has `A:10, B:10, C:10` we will get `AlreadyMerged`, because we ignore C.
But we should not ingore this document and treat it as a "Conflict"