RavenDB-17373 Optimize range queries
Initial support for caching of range queries. At this point, we are simply remembering (per segment reader), the values of a range clause and can reuse that across queries.
At this point, we can see (much) higher performance, as expected, however, we need to also handle query expiry, hooking to the low memory handler, etc.
Another consideration is how to test this.