RavenDB-17241: Support for UTF16 in order to simplify text handling
Given that .Net uses UTF-16 natively all the primitives to deal with text are coded for UTF-16. The complexity to build UTF-8 text handling is outside of the scope of what we should be doing in the short term. The solution is creating a reasonable codepath that can use native UTF8 representation as long as possible but then continue with UTF16 in case there is no choice and we need to transcode the strings to work with. The more native components we support in UTF8 the more we will be able to avoid UTF16 transcoding.