RavenDB-18872: Changing the underlying assumptions for IndexEntryWriter
In the beginning the underlying assumption for the usage of `IndexEntryWriter` was that the caller would provide a big enough buffer to write the object in. That assumption is incorrect, because the caller most likely will not be able to know that beforehand; therefore, now `IndexEntryWriter` will handle internally its' own auxiliary buffer and return the serialized object in a brand new memory segment and provided the caller of `.Finish(out output)` the ability to return that memory when no longer needed.