RavenDB-15895 On large database, index stats are not shown
When we need to estimate the indexing progress on very large indexes, we'll scan only about ~1 million entries to get an exact count.
Afterward, we'll use an estimated value, to reduce the cost of this estimation.

To give some context, if we have 750 million items, we have ~1.47 million pages just for the collection etags). That is ~11GB or so just for that.
We were scanning through all of that on a short duration, leading to both a lot of cost and not actually getting an answer ever.