RavenDB-19278 Fixing faulty error handling during the recovery on storage environment startup. This is related to the introduction of recyclable journals. The main problem was the usage of TryReadAndValidateHeader() method in different mode of journal reading. The added logic was to continue to read and validate the journal to the end after reaching garbage to verify there is no valid later tx which would mean we have true corruption rather then random garbage (https://github.com/ravendb/ravendb/pull/1415). Unfortunatelly this caused that we could shadow an original error and throw errors from the reading continuation while it is expected to get some garbage then.
The fix introduces JournalReader.ReadExpectation enum that allows us to control the journal reader behavior. Now we can have different validation logic depending on the mode that we read in. We do regular validation when we expect to read ValidTransaction but when we encounter some garbage we switch to ZerosOrGarbage reading mode so we can ignore some data errors that we expect due to journals reuse feature.

Important note: we have a concept of "partial recovery" which continues opening a storage environment regardless the encountered data errors (InvalidDataException - e.g a tx has invalid hash, we stop processing further journals then and report about loosing data from some point). When reading the journal in ZerosOrGarbage mode we ignore InvalidDataException but we _don't_ ignore InvalidJournalException which is critical error which we cannot just ignore - it might be that we found that a journal is missing or order of transactions is broken. We need to error if that's the case.