RavenDB-19328 Missing handling of RequireHeaderUpdate when doing partial recovery. We didn't update the header file after detecting that found invalid transaction and we switched to partial recovery mode. Then we have to set JounralInfo.CurrentJournal to the last processed and recovered journal. We have deleted this handling in the following PR: https://github.com/ravendb/ravendb/pull/8556
Another fix here is the erase of remaining part of a partialy recovered journal - transactions that we didn't apply after detecting an error (note that they might have valid hashes). After the partial recovery and load of storage environment, the journal will get new transactions (and ovewrite existing data). New transactions will get the same Ids as the ones which were already written. That can cause problems if we restart the environment again. Then the recovery process might detect transactions that were previously skipped as valid and could throw recovery error. Instead of dealing with detection of this case it's easier to just zero the unprocessed part of partially recovered journal.