Indicate that document is new by providing an empty change vector fallback in ConcurrencyCheckMode.Forced scenario
This has no functional impact as DocumentPutAction.PutAction method doesn't differentiate at the moment between null or empty change vector when there was no such document before on the server side. If in the future there was a need to differentiate then this change would be needed anyway. But for now the only point of this change is to bring consistency with ConcurrencyCheckMode.Auto scenario to avoid confusion of why ConcurrencyCheckMode.Auto fallbacks to empty change vector and ConcurrencyCheckMode.Forced does not.