RavenDB-19992: Alternative implementation of the PersistanceDictionaryLocator
After talking with @ayende we agreed that given the idea of the locator is to try to avoid as much the code complexity of having to deal with the collisions with complex codepaths. We decided to settle for a suboptimal solution but continue with it until it becomes an actual performance issue.