RavenDB-22457
Ensuring that we'll keep the reference for the old state of the compression buffer for the duration of PrepareToWriteToJournal.
This is required because are making *two* calls to EnsureContinious, and may create a new mapping because of that, and the old state (and mapping) may be cleared away by the finalizer *while the method is running.