RavenDB-20877: Cached properties can survive between transaction
The after effect is that we don't need to pay for disposing because the probability of dealing with similar properties is high and there is already mechanisms in place to deal with the case when that is not the case. Most of the worked performed on Dispose was actually redundant and hampered the optimization over high traffic burst of the same database and collection.