RavenDB-24020 - Changing HNSW insert to parallel mode.
It is now possible to interleave the IO work of multiple nodes in the graph at the same time that we are doing distance computation in the thread pool. We do that by interleaving (using yield) multiple "concurrent" node placement routines that
share the same runner. The runner runs each of them in turn and amortize all access to the disk (to load vectors) and then execute the actual distance computation on the thread pool. The runner is in charge of this whole orchestration, with the
idea being that we can run both computation and disk work at the same time.
This is critical because if we have a graph to insert that is larger than memory, we may be stall by swapping for basically forever without this approach.

- Because we are now doing multiple in-flight nodes, we add the previous N (number of edges) nodes as presumptive edges to the current node, to "kick start" the graph wiring process. This ends up

- Using the .NET default thread pool because all attempts (blocking collection, channels, dedicated threads, etc) has ended up significantly worse in terms of performance.
  The design means that we have a max of N batches of work to do (where N is likely small and fixed) and those tend to be very short lived. So even under load, it isn't expected to fill the thread pool and cause a train.

- Add max batch count to allow user to control the concurrency level
- Adding debug endpoint to expose the internal structure of HNSW index
- Fixing generation of debug output for HNSW
- Allow to generate debug output on machines without chrome